refactor : nested record로 변경 및 League 변환용 메서드 추가
